In courseDetails model courseId datatype is changed long to int

+++ src/main/java/com/codegnan/model/CourseDetails.java
@@ -65,7 +65,7 @@ public class CourseDetails {
 	 * @param instructorName
 	 * @param studentsCourses
 	 */
-	public CourseDetails(long courseId, String courseName, String courseDiscription, double coursePrice,
+	public CourseDetails(int courseId, String courseName, String courseDiscription, double coursePrice,
 			String instructorName, List<String> studentsCourses) {
 		super();
 		this.courseId = courseId;
